Gestion de Plan de Pagos 2/2

+++ resources/views/pago/index.blade.php
@@ -41,23 +41,47 @@
 									<th>Saldo</th>
 									<th>Proxima Cita</th>
 									@if (sizeof($pagos) == 0)
-									<th width="237">
-										@can('pagos.create')
-										<a href="{{ route('pago.create',['cita' => $id]) }}" class="btn btn-block btn-success pull-right">
-										<i class="fa fa-money"></i>	
-										Crear Pago
-										</a>
-										@endcan
-									</th>
+										@if($abonoValidar == 0 )
+
+										@else
+											@if($saldo == 0)
+											<th></th>
+											@else
+												@if(sizeof($pagos) > 1)
+												
+												@else
+													<th width="237">
+														@can('pagos.create')
+														<a href="{{ route('pago.create',['cita' => $id]) }}" class="btn btn-block btn-success pull-right">
+														<i class="fa fa-money"></i>	
+														Crear Pago
+														</a>
+														@endcan
+													</th>
+												@endif
+											@endif
+										@endif
 									@else
-									<th width="237">
-										@can('pagos.create')
-										<a href="{{ route('pago.create', ['cita' => $id]) }}" class="btn btn-block btn-success pull-right">
-										<i class="fa fa-money"></i>	
-										Crear Pago
-										</a>
-										@endcan
-									</th>
+										@if($abonoValidar == 0)
+										
+										@else
+											@if($saldo == 0)
+											<th></th>
+											@else
+												@if(sizeof($pagos) > 1)
+													<th width="237">
+														@can('pagos.create')
+														<a href="{{ route('pago.create', ['cita' => $id]) }}" class="btn btn-block btn-success pull-right">
+														<i class="fa fa-money"></i>	
+														Crear Pago
+														</a>
+														@endcan
+													</th>
+												@else
+												<th></th>
+												@endif
+											@endif
+										@endif
 									@endif
 								</tr>
 							</thead>
@@ -78,38 +102,42 @@
 										$aux= date_format($date,"d-m-Y");
 									@endphp
 									<td align="center">{{$aux}}</td>
-									<td width="10px">
-										@can('pagos.destroy')
-											<button type="button" class="btn btn-sm btn-default btn btn-danger" data-toggle="modal" data-target="#Modal{{$proceso->id}}" style="width: 100%"><i class="fa fa-trash"></i>
-													  					Eliminar
-																	</button>
-																	{!! Form::open(['route' => ['pago.destroy', $proceso->id],'method' => 'DELETE']) !!}
-																		<!-- Modal -->
-																		<div class="modal fade" id="Modal{{$proceso->id}}" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
-																			<div class="modal-dialog" role="document">
-																		    	<div class="modal-content">
-																		    		<div class="modal-header">
-																		        		<h5 class="modal-title" id="exampleModalLabel">Eliminar Pago</h5>
-																		        			<button type="button" class="close" data-dismiss="modal" aria-label="Close">
-																		          				<span aria-hidden="true">&times;</span>
-																		        			</button>
-																		      		</div>
-																		      		<div class="modal-body">
-																		      			<label>Estas seguro?</label>
-																		      			<br/>
-																		      			<button type="button" class="btn btn-md btn-default" data-dismiss="modal">No</button>
-																		        		<button class="btn btn-md btn-default bg-danger" style="color: white">
-																							Si
-																						</button>
-																		      		</div>
-																		      		<div class="modal-footer">
-																		      		</div>
-																		    	</div>
+									@if($abonoValidar != 0)
+										<td width="10px">
+											@can('pagos.destroy')
+												<button type="button" class="btn btn-sm btn-default btn btn-danger" data-toggle="modal" data-target="#Modal{{$proceso->id}}" style="width: 100%"><i class="fa fa-trash"></i>
+														  					Eliminar
+																		</button>
+																		{!! Form::open(['route' => ['pago.destroy', $proceso->id],'method' => 'DELETE']) !!}
+																			<!-- Modal -->
+																			<div class="modal fade" id="Modal{{$proceso->id}}" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
+																				<div class="modal-dialog" role="document">
+																			    	<div class="modal-content">
+																			    		<div class="modal-header">
+																			        		<h5 class="modal-title" id="exampleModalLabel">Eliminar Pago</h5>
+																			        			<button type="button" class="close" data-dismiss="modal" aria-label="Close">
+																			          				<span aria-hidden="true">&times;</span>
+																			        			</button>
+																			      		</div>
+																			      		<div class="modal-body">
+																			      			<label>Estas seguro?</label>
+																			      			<br/>
+																			      			<button type="button" class="btn btn-md btn-default" data-dismiss="modal">No</button>
+																			        		<button class="btn btn-md btn-default bg-danger" style="color: white">
+																								Si
+																							</button>
+																			      		</div>
+																			      		<div class="modal-footer">
+																			      		</div>
+																			    	</div>
+																				</div>
 																			</div>
-																		</div>
-																	{!! Form::close() !!}
-										@endcan
-									</td>
+																		{!! Form::close() !!}
+											@endcan
+										</td>
+										@else
+										<td></td>
+										@endif
 								</tr>
 								@endforeach
 							</tbody>
